Fix and simplify user avatar loading
The old code did several things we don't want or need anymore.
1) Lots of cairo drawing to achieve a rounded rectangle as border for
the user avatar.
2) Lots of file checks, which we don't really need as
gdk_pixbuf_new_from_file gives us a proper error message if the image
cannot be loaded. The old code contained no workarounds for when an
image file didn't comply, so we may as well rely on gdk to decide.