Add support for elogind (bug #15243)
Add support for elogind, the split-off implementation of systemd-logind
for non-systemd systems.  It is mostly API-compatible, with the notable
exception of renaming the library, so I've managed to keep changes
to the minimum.  I've replaced 'systemd' and 'SYSTEMD_LOGIND' with plain
'logind' to avoid the confusion that they are only used by systemd
branches of the code.

Signed-off-by: Michał Górny <mgorny@gentoo.org>
Signed-off-by: Sean Davis <smd.seandavis@gmail.com>