Fix inhibitor listing in xfce4-screensaver-command (bug #16355)
The command would not reflect inhibited status and would not show the
inhibitors. This was due to the fact that unref'ing the reply gobject
also invalidates the body extracted from it and the iterator was set up
with the wrong format string.

Also, the body is a container and always contains the inhibitor array
element, even if it's empty. Therefore we have to check the number of
children of the actual array variant and not the body.

Change both to get inhibitors listed and inhibition status shown when
querying status.

Signed-off-by: Sean Davis <smd.seandavis@gmail.com>