Fix inhibitor proxying (bug #16356)
When inhibiting the screensaver using e.g. xfce4-screensaver-command -q,
xfce4-screensaver would report an error:

org.freedesktop.DBus.Error.ServiceUnknown raised: The name
org.freedesktop.login1 was not provided by any .service files

Consequently, no inhibitor was added to logind. This was due
to the fact that the wrong connection was used for the dbus message and
thus logind expected on the session bus.

Switching to the system connection to correctly relay the inhibitor to
logind.

Signed-off-by: Sean Davis <smd.seandavis@gmail.com>