dbus: Prevent overzealous activiation (bug #16365)
Port over a check from mate screensaver that prevents activation when
clearing an inhibitor by first checking if the session is actually idle.
If clearing the inhibitor is the result of some interactive user action,
this should never be the case and thus prevent e.g. saver activation
when stopping video playback that did inhibit the screen saver.

Signed-off-by: Sean Davis <smd.seandavis@gmail.com>