Round the user picture on lock dialog.
All user dialogs (Linux, Windows, Mac, Android, etc..), or any
representation of a user as an avatar in the web are round
images.

Also, it shows an emblem icon to note that it is logged in as
in other dialogs.

Both options can be made configurable by the user very easily,
but at least (IMHO of course), I think it respects the
preference of the majority of users.