savers: Fix critical warnings from GSThemeWindow on "realize"
A minimal fix that lets GTK create the widget window and then set the
remote window as its parent, rather than forcing its use as widget
window. There may be some unnecessary things to remove after this change
and some cleanup to do more generally, but the idea is to avoid side
effects as much as possible for now.