Fix build on FreeBSD
On FreeBSD environ needs to be defined in the same source file where it is used.

Linux provides a nonstandard extention to avoid needing the definition.

This patch adds back code which was removed in commit 447b31948f75d133cd33e3ff3620e92d372e68ef but with #ifdef __FreeBSD__ since it is needed only there.

Ref: https://man7.org/linux/man-pages/man7/environ.7.html
     https://man.freebsd.org/cgi/man.cgi?query=environ&apropos=0&sektion=0&manpath=FreeBSD+13.1-RELEASE+and+Ports&arch=default&format=html