screensaver-configure: Fix broken translations
This was broken because the last argument of os.path.join() must not
start with `/` otherwise it is interpreted as an absolute path and all
previous arguments are ignored.

This change also makes the locale directory dependent on the
installation based on where the script is located (i.e. normally where
it was installed). This should work in general, although it might be
cleaner to do this via autotools with a `.in` file.

Finally, there is no need to add Libxfce4util.textdomain() everywhere,
only once in main() is enough.