screensaver-configure: Ensure consistent dialog positioning
If the position of the main dialog is left at NONE, depending on its
content the window manager may decide to place it in different places,
which results in inconsistent behavior (e.g. for me the blank screen
dialog is positioned at the top left of the screen while the others are
positioned under the mouse).

Also, since it is not possible to use set_transient_for() here,
positioning under the mouse seems to be the best choice to emulate this
behavior.