Use GtkSocket/GtkPlug instead of GtkDrawingArea/GSThemeWindow
The precautions taken in this commit to avoid X11 errors are not always
sufficient. For example, it's possible to get one of these errors on
resuming from suspend when a notification appears at the same time (not
just any notification, apparently; in my case it's from nm-applet).
However, I can't reproduce the problem with openbox, only with xfwm4,
and apparently only with /general/unredirect_overlays enabled.