wayland: Make sure text entry of unlock dialog is grabbed
The ext-session-lock protocol imposes nothing on the compositor
concerning the focus of lock surfaces. In particular, there is no
guarantee that the user will be able to click on a surface to give it
focus, even if the protocol suggests this.

More generally, it's not certain that we receive pointer events on lock
surfaces, and even when we do, we don't necessarily have a reliable way
of knowing which surface is under the pointer at any given time.

At first glance, then, it seems that the only thing we can do to ensure
that the unlock dialog has indeed the focus, is to display it on the
lock surface to which the compositor has given the focus when creating
all the surfaces. This is what the Wayland implementation of
find_window_at_pointer() does in this commit.

However, the active surface may have changed by the time the dialog is
displayed, particularly when user activity takes place just after
surfaces have been created. In this case, we cancel the dialog display,
and another user activity will re-trigger its creation (in practice,
this simply adds a small latency).

Next, we need to add a focus recovery timeout at plug level, in case
focus is lost when exiting text entry, for example by pressing the tab
key (for some reason, pressing this key doesn't automatically return you
to text entry once you've reached the last button in the action area, as
on X11).