xfce4-screensaver-dialog: Remove custom status text mapper
This is motivated by the previous commit ('xfce4-screensaver-dialog:
Hide translated "Password:" message').

When the strings are translated in pam this mapper won't work. So it
cannot really "change some of these messages to be more sane" in most
case.

Also the strings here are out-of-sync with pam, for example in
linux-pam@0fb1f5b7 trailing periods are added to many strings, breaking
the mapping completely.

https://github.com/linux-pam/linux-pam/commit/0fb1f5b701432b6d9f40754b69f2fe0dc6f75103