manager: Use only one permanent window as overlay
Given the asynchronous nature of X11, this seems to be the only way to
ensure that the entire desktop will be covered at all times. Also, this
is something that is only useful on X11 (on Wayland, the
ext-session-lock protocol is specifically designed to avoid all these
problems), so it's best to restrict it to X11.

Fixes: #138